finish SIR for agents
