+++ src/SIR.cpp
@@ -1,7 +1,7 @@
 /****************
  * COVID-19ABMGuelphS20
- * 30/06/20
- * ver 0.01
+ * 14/07/20
+ * ver 0.02
  * 
  * This is the class file for the SIR class
  ***************/
@@ -15,22 +15,27 @@ SIR::SIR() {
     currentSeverity = SUSCEPTIBLE;
 }
 
+// leave these
 // check all references
 // can all be changed to user inputs
 // 45% asymptomatic if not srs (check)
-// 25% icu death rate
-// approx 20% need hospital
-// 6% need ICU
-// case peak around 10 days
-// around 14 days to death (check)
-// around 26 days to srs recovery via icu
+// 25% icu death rate once in ICU
+// approx 20% need hospital, 80% recover at home
+// 6% need ICU of total cases
+// around 26 days to srs recovery once in icu
 // icu 1 to 3 days after hospitilization
-// icu recovery 14 days
+// non-icu recovery 14 days
 // icu death after 5 days
-// days til icu reocvery 14 days
 
+//possible timestep 0.125 3 hrs
+
+/*************************
+ * SIRTimeStep
+ * 
+ * This progresses agents through their various SIR model paths.
+ ************************/
 void SIR::SIRTimeStep(double timeStep) {
-    // incubating stage
+    // incubating stage infect other ppl?
     if (isIncubating) {
         incubationPeriod -= timeStep;
 
@@ -64,6 +69,11 @@ void SIR::SIRTimeStep(double timeStep) {
     }
 }
 
+/*************************
+ * DecideSIRCase
+ * 
+ * Determines the path the agents COVID19 case will follow.
+ ************************/
 void SIR::DecideSIRCase(double infectedNumb, double infectedChance) {
     // decide if the case is fatal
     if (infectedNumb <  infectedChance) {
@@ -76,27 +86,31 @@ void SIR::DecideSIRCase(double infectedNumb, double infectedChance) {
         showsSymptoms = true;
       // if the case is not fatal decide path to recovery
     } else {
-        double hospitalAsymptomRisk = (double) rand()/RAND_MAX;
+        double randomNumber = (double) rand()/RAND_MAX;
         fatalCase = false;
         // case is asymptomatic
-        if (hospitalAsymptomRisk < 0.45) { // asymptomatic probability if non fatal (user input)
+        if (randomNumber < 0.45) { // asymptomatic probability if non fatal (user input)
             showsSymptoms = false;
             needHospital = false;
             needIcu = false;
             timeTilRecovery = 14; // asymptomatic recovery (user input)
-        } else if (0.45 <= hospitalAsymptomRisk <= 0.65) { //hospitilization chance per case (user input)
+          // case requires hospitilzation
+        } else if (0.45 <= randomNumber <= 0.65) { //hospitilization chance per case (user input)
             needHospital = true;
             showsSymptoms = true;
             timeTilHospital = 7; //time needed to get srs enough for hospital (make user input)
             double icuNeeded = (double) rand()/RAND_MAX;
+            // case requires icu
             if (icuNeeded < 0.06) { //chance of u needing icu if hospitalized (look this up)
                 timeTilICU = 3; //time needed from hospital to ICU (make user input)
                 timeTilRecovery = 26; //recovery from ICU on average (make user input)
                 needIcu = true;
+              // case requires general ward hospital
             } else {
                 needIcu = false;
-                timeTilRecovery = 12; //Time til Hospital recovery (make user input)
+                timeTilRecovery = 10; //Time til Hospital recovery (make user input)
             }
+          // case requires no hospital but shows symptoms
         } else {
             showsSymptoms = true;
             needHospital = false;
@@ -106,46 +120,57 @@ void SIR::DecideSIRCase(double infectedNumb, double infectedChance) {
     }
 }
 
-// Infect the agent and decide their path through the SIR model
+/*************************
+ * AgentInfected
+ * 
+ * Infect the agent and decide their path through the SIR model.
+ ************************/
 void SIR::AgentInfected(AgentInfo info) {
     currentSeverity = INFECTED;
     isIncubating = true;
-    incubationPeriod = std::rand()%7 + std::rand()%7;
+    incubationPeriod = std::rand()%7 + std::rand()%7 + 1;
 
-    double randomNumber = (double) rand()/RAND_MAX;
+    double randomDeathSentence = (double) rand()/RAND_MAX;
 
-    // based on worldometer death rates
+    // based on worldometer death rates for each age group
     if(MALE0TO4 == info || MALE5TO9 == info || FEMALE0TO4 == info || FEMALE5TO9 == info) {
-        DecideSIRCase(randomNumber, 0.0);
+        DecideSIRCase(randomDeathSentence, 0.0);
     } else if (MALE10TO14 == info || MALE15TO19 == info || FEMALE10TO14 == info || FEMALE15TO19 == info) {
-        DecideSIRCase(randomNumber, 0.002);
+        DecideSIRCase(randomDeathSentence, 0.002);
     } else if (MALE20TO24 == info || MALE25TO29 == info || FEMALE20TO24 == info || FEMALE25TO29 == info) {
-        DecideSIRCase(randomNumber, 0.002);
+        DecideSIRCase(randomDeathSentence, 0.002);
     } else if (MALE30TO34 == info || MALE35TO39 == info || FEMALE30TO34 == info || FEMALE35TO39 == info) {
-        DecideSIRCase(randomNumber, 0.002);
+        DecideSIRCase(randomDeathSentence, 0.002);
     } else if (MALE40TO44 == info || MALE45TO49 == info || FEMALE40TO44 == info || FEMALE45TO49 == info) {
-        DecideSIRCase(randomNumber, 0.004);
+        DecideSIRCase(randomDeathSentence, 0.004);
     } else if (MALE50TO54 == info || MALE55TO59 == info || FEMALE50TO54 == info || FEMALE55TO59 == info) {
-        DecideSIRCase(randomNumber, 0.013);
+        DecideSIRCase(randomDeathSentence, 0.013);
     } else if (MALE60TO64 == info || MALE65TO69 == info || FEMALE60TO64 == info || FEMALE65TO69 == info) {
-        DecideSIRCase(randomNumber, 0.036);
+        DecideSIRCase(randomDeathSentence, 0.036);
     } else if (MALE70TO74 == info || MALE75TO79 == info || FEMALE70TO74 == info || FEMALE75TO79 == info) {
-        DecideSIRCase(randomNumber, 0.08);
+        DecideSIRCase(randomDeathSentence, 0.08);
     } else {
-        DecideSIRCase(randomNumber, 0.148);
+        DecideSIRCase(randomDeathSentence, 0.148);
     }
 }
 
+/*************************
+ * The functions below move the agent through the SIR model.
+ ************************/
+
 void SIR::HospitalAgent() {
     currentSeverity = HOSPITALIZED;
+    needHospital = false;
 }
 
 void SIR::QuarantineAgent() {
     currentSeverity = ISOLATED;
+    isIncubating = false;
 }
 
 void SIR::PlaceAgentInICU() {
     currentSeverity = ICU;
+    needIcu = false;
 }
 
 void SIR::RecoverAgent() {
@@ -155,6 +180,12 @@ void SIR::RecoverAgent() {
 void SIR::AgentDeceased() {
     currentSeverity = DECEASED;
 } 
+
+/*************************
+ * DetermineSeverity
+ * 
+ * Get the current severity of the agents case.
+ ************************/
 SIRSeverity SIR::DetermineSeverity() {
     return currentSeverity;
 }
